# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MinuteOps API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools:
    get:
      operationId: listSchools
      tags: [students]
      summary: List schools
      responses:
        "200":
          description: List of schools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
    post:
      operationId: createSchool
      tags: [students]
      summary: Create school
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolBody"
      responses:
        "201":
          description: Created school
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /programs:
    get:
      operationId: listPrograms
      tags: [students]
      summary: List programs
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [students]
      summary: Create program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"

  # ─── Students ────────────────────────────────────────────────────
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff
      parameters:
        - name: role
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /staff/{id}:
    get:
      operationId: getStaff
      tags: [staff]
      summary: Get staff detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffDetail"
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /staff/{id}/caseload:
    get:
      operationId: getStaffCaseload
      tags: [staff]
      summary: Get caseload and minute progress for a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Caseload
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MinuteProgress"

  # ─── Service Types ────────────────────────────────────────────────
  /service-types:
    get:
      operationId: listServiceTypes
      tags: [services]
      summary: List service types
      responses:
        "200":
          description: Service types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceType"
    post:
      operationId: createServiceType
      tags: [services]
      summary: Create service type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceTypeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceType"

  # ─── Service Requirements ─────────────────────────────────────────
  /service-requirements:
    get:
      operationId: listServiceRequirements
      tags: [services]
      summary: List service requirements
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: serviceTypeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: providerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: active
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Requirements list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequirement"
    post:
      operationId: createServiceRequirement
      tags: [services]
      summary: Create service requirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequirementBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequirement"
  /service-requirements/{id}:
    get:
      operationId: getServiceRequirement
      tags: [services]
      summary: Get service requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequirement"
    patch:
      operationId: updateServiceRequirement
      tags: [services]
      summary: Update service requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequirementBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequirement"
    delete:
      operationId: deleteServiceRequirement
      tags: [services]
      summary: Delete service requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Session Logs ────────────────────────────────────────────────
...
components:
  schemas:
...
    School:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        district:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, createdAt]
    CreateSchoolBody:
      type: object
      properties:
        name:
          type: string
        district:
          type: ["string", "null"]
      required: [name]
    Program:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        schoolId:
          type: integer
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, schoolId, createdAt]
    CreateProgramBody:
      type: object
      properties:
        name:
          type: string
        schoolId:
          type: integer
        description:
          type: ["string", "null"]
      required: [name, schoolId]

    # ─── Students ─────────────────────────────────────────────────
...
    Staff:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        title:
          type: ["string", "null"]
        schoolId:
          type: ["integer", "null"]
        status:
          type: string
        qualifications:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, firstName, lastName, role, status, createdAt]
    StaffDetail:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        title:
          type: ["string", "null"]
        schoolId:
          type: ["integer", "null"]
        status:
          type: string
        qualifications:
          type: ["string", "null"]
        createdAt:
          type: string
        assignedStudents:
          type: array
          items:
            $ref: "#/components/schemas/StaffAssignment"
        scheduleBlocks:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleBlock"
      required: [id, firstName, lastName, role, status, createdAt, assignedStudents, scheduleBlocks]
    CreateStaffBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        title:
          type: ["string", "null"]
        schoolId:
          type: ["integer", "null"]
        status:
          type: string
        qualifications:
          type: ["string", "null"]
      required: [firstName, lastName, role, status]
    UpdateStaffBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        schoolId:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
        qualifications:
          type: ["string", "null"]

    # ─── Service Types ────────────────────────────────────────────
    ServiceType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        color:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        defaultIntervalType:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, category, createdAt]
    CreateServiceTypeBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        color:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        defaultIntervalType:
          type: ["string", "null"]
      required: [name, category]

    # ─── Service Requirements ─────────────────────────────────────
    ServiceRequirement:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        serviceTypeId:
          type: integer
        serviceTypeName:
          type: ["string", "null"]
        providerId:
          type: ["integer", "null"]
        providerName:
          type: ["string", "null"]
        deliveryType:
          type: string
        requiredMinutes:
          type: integer
        intervalType:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
      required: [id, studentId, serviceTypeId, deliveryType, requiredMinutes, intervalType, startDate, active, createdAt]
    CreateServiceRequirementBody:
      type: object
      properties:
        studentId:
          type: integer
        serviceTypeId:
          type: integer
        providerId:
          type: ["integer", "null"]
        deliveryType:
          type: string
        requiredMinutes:
          type: integer
        intervalType:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
      required: [studentId, serviceTypeId, deliveryType, requiredMinutes, intervalType, startDate, active]
    UpdateServiceRequirementBody:
      type: object
      properties:
        providerId:
          type: ["integer", "null"]
        deliveryType:
          type: ["string", "null"]
        requiredMinutes:
          type: ["integer", "null"]
        intervalType:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]

    # ─── Session Logs ─────────────────────────────────────────────
...
    MinuteProgress:
      type: object
      properties:
        serviceRequirementId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        serviceTypeId:
          type: integer
        serviceTypeName:
          type: string
        providerId:
          type: ["integer", "null"]
        providerName:
          type: ["string", "null"]
        intervalType:
          type: string
        requiredMinutes:
          type: integer
        deliveredMinutes:
          type: integer
        remainingMinutes:
          type: integer
        percentComplete:
          type: number
        expectedMinutesByNow:
          type: number
        projectedMinutes:
          type: number
        riskStatus:
          type: string
        intervalStart:
          type: string
        intervalEnd:
          type: string
        missedSessionsCount:
          type: integer
        makeupSessionsCount:
          type: integer
      required: [serviceRequirementId, studentId, studentName, serviceTypeId, serviceTypeName, intervalType, requiredMinutes, deliveredMinutes, remainingMinutes, percentComplete, expectedMinutesByNow, projectedMinutes, riskStatus, intervalStart, intervalEnd, missedSessionsCount, makeupSessionsCount]

    # ─── Schedule Blocks ─────────────────────────────────────────
    ScheduleBlock:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        studentId:
          type: ["integer", "null"]
        studentName:
          type: ["string", "null"]
        serviceTypeId:
          type: ["integer", "null"]
        serviceTypeName:
          type: ["string", "null"]
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: ["string", "null"]
        blockLabel:
          type: ["string", "null"]
        blockType:
          type: string
        notes:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        weekOf:
          type: ["string", "null"]
        isAutoGenerated:
          type: boolean
        createdAt:
          type: string
      required: [id, staffId, dayOfWeek, startTime, endTime, blockType, isRecurring, isAutoGenerated, createdAt]
...
    StaffAssignment:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        staffRole:
          type: ["string", "null"]
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        assignmentType:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, staffId, studentId, assignmentType, createdAt]
...