# lib/api-zod/src/generated/types/createServiceTypeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MinuteOps API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateServiceTypeBody {
  name: string;
  category: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  defaultIntervalType?: string | null;
}